+++ lib/components/floating_island.dart
@@ -236,13 +236,14 @@ class _CNFloatingIslandState extends State<CNFloatingIsland>
     final route = ModalRoute.of(context);
     final newAnim = route?.secondaryAnimation;
     if (identical(newAnim, _secondaryRouteAnim)) return;
-    _secondaryRouteAnim?.removeListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.removeStatusListener(_onSecondaryRouteAnimChanged);
     _secondaryRouteAnim = newAnim;
-    _secondaryRouteAnim?.addListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.addStatusListener(_onSecondaryRouteAnimChanged);
     _onSecondaryRouteAnimChanged();
   }
 
-  void _onSecondaryRouteAnimChanged() => _pushContainmentIfNeeded();
+  void _onSecondaryRouteAnimChanged([AnimationStatus? status]) =>
+      _pushContainmentIfNeeded();
 
   void _pushContainmentIfNeeded() {
     final anim = _secondaryRouteAnim;
@@ -278,7 +279,7 @@ class _CNFloatingIslandState extends State<CNFloatingIsland>
 
   @override
   void dispose() {
-    _secondaryRouteAnim?.removeListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.removeStatusListener(_onSecondaryRouteAnimChanged);
     _secondaryRouteAnim = null;
     _animationController.dispose();
     _controller._detach();