+++ lib/components/glass_button_group.dart
@@ -152,7 +152,7 @@ class _CNGlassButtonGroupState extends State<CNGlassButtonGroup>
 
   @override
   void dispose() {
-    _secondaryRouteAnim?.removeListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.removeStatusListener(_onSecondaryRouteAnimChanged);
     _secondaryRouteAnim = null;
     super.dispose();
   }
@@ -161,13 +161,14 @@ class _CNGlassButtonGroupState extends State<CNGlassButtonGroup>
     final route = ModalRoute.of(context);
     final newAnim = route?.secondaryAnimation;
     if (identical(newAnim, _secondaryRouteAnim)) return;
-    _secondaryRouteAnim?.removeListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.removeStatusListener(_onSecondaryRouteAnimChanged);
     _secondaryRouteAnim = newAnim;
-    _secondaryRouteAnim?.addListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.addStatusListener(_onSecondaryRouteAnimChanged);
     _onSecondaryRouteAnimChanged();
   }
 
-  void _onSecondaryRouteAnimChanged() => _pushContainmentIfNeeded();
+  void _onSecondaryRouteAnimChanged([AnimationStatus? status]) =>
+      _pushContainmentIfNeeded();
 
   void _pushContainmentIfNeeded() {
     final anim = _secondaryRouteAnim;