+++ lib/components/liquid_glass_container.dart
@@ -87,7 +87,7 @@ class _LiquidGlassContainerState extends State<LiquidGlassContainer>
 
   @override
   void dispose() {
-    _secondaryRouteAnim?.removeListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.removeStatusListener(_onSecondaryRouteAnimChanged);
     _secondaryRouteAnim = null;
     PlatformViewGuard.readyNotifier.removeListener(_onPlatformViewGuardReady);
     _channel?.setMethodCallHandler(null);
@@ -99,13 +99,14 @@ class _LiquidGlassContainerState extends State<LiquidGlassContainer>
     final route = ModalRoute.of(context);
     final newAnim = route?.secondaryAnimation;
     if (identical(newAnim, _secondaryRouteAnim)) return;
-    _secondaryRouteAnim?.removeListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.removeStatusListener(_onSecondaryRouteAnimChanged);
     _secondaryRouteAnim = newAnim;
-    _secondaryRouteAnim?.addListener(_onSecondaryRouteAnimChanged);
+    _secondaryRouteAnim?.addStatusListener(_onSecondaryRouteAnimChanged);
     _onSecondaryRouteAnimChanged();
   }
 
-  void _onSecondaryRouteAnimChanged() => _pushContainmentIfNeeded();
+  void _onSecondaryRouteAnimChanged([AnimationStatus? status]) =>
+      _pushContainmentIfNeeded();
 
   void _pushContainmentIfNeeded() {
     final anim = _secondaryRouteAnim;